Almost done... need to test, implement debug messages, and finish game logic.

+++ HW_6/p2/sem.c
@@ -4,11 +4,16 @@
 
 #include <signal.h>
 #include <stdlib.h>
+#include <unistd.h>
 #include "spinlock.h"
 
 void sem_init(struct sem *s, int count) {
     s->count = count;
     s->proc_count = 0;
+    //Not sure if this is even necessary as mmap guarantees empty memory
+    for(int i = 0; i < MAX_PROCS; i++){
+        s->IDs[i] = 0;
+    }
     spin_unlock(s->lock);
 }
 
@@ -23,7 +28,7 @@ int sem_try(struct sem *s) {
     return 1;
 }
 
-void sem_wait(struct sem *s, int id) {
+void sem_wait(struct sem *s, int vproc) {
     sigset_t full_mask;
     sigset_t suspend_mask;
     sigemptyset(&full_mask);
@@ -33,10 +38,12 @@ void sem_wait(struct sem *s, int id) {
     while (sem_try(s) != 0) {
         spin_lock(s->lock);
         sigprocmask(SIG_BLOCK, &full_mask, NULL);
-        s->IDs[s->proc_count] = id;
+        s->IDs[vproc] = getpid();
+        s->instrumentation[vproc].sleep++;
         s->proc_count++;
         spin_unlock(s->lock);
         sigsuspend(&suspend_mask);
+        s->instrumentation[vproc].wake++;
     }
 }
 
@@ -44,8 +51,10 @@ void sem_inc(struct sem *s){
     spin_lock(s->lock);
     s->count++;
     if(s->count > 0){
-        for(int i = 0; i < s->proc_count; i++){
-            kill(s->IDs[i],SIGUSR1);
+        for(int i = 0; i < MAX_PROCS; i++){
+            if(s->IDs[i] == 1){
+                kill(s->IDs[i],SIGUSR1);
+            }
         }
     }
     spin_unlock(s->lock);
